# interface/docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact: {}
  description: A simple file upload API with RabbitMQ and Python worker
  title: File Upload API
  version: "1.0"
paths:
  /upload:
    post:
      consumes:
        - multipart/form-data
      description: Accepts a file and sends its name to RabbitMQ
      parameters:
        - description: File to upload
          in: formData
          name: file
          required: true
          type: file
      produces:
        - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
      summary: Uploads a file
      tags:
        - files
  /predict:
    post:
      consumes:
        - application/json
      description: Predicts the price of a car based on its attributes
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              brand:
                type: string
              year:
                type: integer
              engine_size:
                type: number
                format: float
              fuel_type:
                type: string
              transmission:
                type: string
              mileage:
                type: integer
              condition:
                type: string
              model:
                type: string
            required:
              - brand
              - year
              - engine_size
              - fuel_type
              - transmission
              - mileage
              - condition
              - model
      produces:
        - application/json
      responses:
        "200":
          description: Prediction result
          schema:
            type: object
            properties:
              price:
                type: number
                format: float
        "400":
          description: Bad request
          schema:
            type: string
      summary: Predicts car price
      tags:
        - predictions

swagger: "2.0"